+++ java-test/src/test/java/io/jeyong/test/environment/test/mode/unit/ExceptionModeUnitTest.java
@@ -18,7 +18,7 @@ class ExceptionModeUnitTest {
     @Autowired
     private AddressRepository addressRepository;
 
-    @Disabled("AfterEachCallback을 통해 발생하는 예외는 테스트할 수 없다.")
+    @Disabled("TestExecutionListener를 통해 발생하는 예외는 테스트할 수 없다.")
     @Test
     @DisplayName("@DataJpaTest를 이용한 Repository를 호출하는 상황에서 EXCEPTION 모드가 동작한다.")
     void testExceptionModeInRepositoryCall() {
