+++ kotlin-test/src/test/kotlin/io/jeyong/test/environment/test/unit/ExceptionModeUnitTest.kt
@@ -18,7 +18,7 @@ class ExceptionModeUnitTest {
     @Autowired
     private lateinit var productRepository: ProductRepository
 
-    @Disabled("AfterEachCallback을 통해 발생하는 예외는 테스트할 수 없다.")
+    @Disabled("TestExecutionListener를 통해 발생하는 예외는 테스트할 수 없다.")
     @Test
     @DisplayName("@DataJpaTest를 이용한 Repository를 호출하는 상황에서 EXCEPTION 모드가 동작한다.")
     fun testExceptionModeInRepositoryCall() {
